+++ src/python_anesthesia_simulator/__init__.py
@@ -1,7 +1,8 @@
-from .simulator import Patient
+from .patient import Patient
 from .disturbances import compute_disturbances
 from .metrics import compute_control_metrics
 from .pk_models import CompartmentModel
 from .pd_models import BIS_model, TOL_model, Hemo_meca_PD_model
 from .tci_control import TCIController
 from .alarms import standard_alarm
+from .simulator import Simulator
