+++ tests/test_simulation.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from python_anesthesia_simulator import simulator
+from python_anesthesia_simulator import patient
 
 
 # Simulation duration in seconds
@@ -36,16 +36,16 @@ remifentanil_infusion_profile[int(200/ts):] = 0.1    # 0.1 ug/s from 200s onward
 norepinephrine_infusion_profile[int(1800/ts):] = 0.1    # 0.1 ug/s from 1800s onward
 
 # Patient object
-George_1 = simulator.Patient([age, height, weight, gender],
-                             ts=ts,
-                             model_propo="Schnider",
-                             model_remi="Minto",
-                             model_nore="Beloeil",
-                             random_PD=False)
+George_1 = patient.Patient([age, height, weight, gender],
+                           ts=ts,
+                           model_propo="Schnider",
+                           model_remi="Minto",
+                           model_nore="Beloeil",
+                           random_PD=False)
 
-df_George_1 = George_1.full_sim(u_propo = propofol_infusion_profile, 
-                                u_remi = remifentanil_infusion_profile,
-                                u_nore = norepinephrine_infusion_profile)
+df_George_1 = George_1.full_sim(u_propo=propofol_infusion_profile,
+                                u_remi=remifentanil_infusion_profile,
+                                u_nore=norepinephrine_infusion_profile)
 
 # %% Simulation ts 50
 # Sampling time in seconds
@@ -71,18 +71,17 @@ remifentanil_infusion_profile[int(200/ts):] = 0.1    # 0.1 ug/s from 200s onward
 norepinephrine_infusion_profile[int(1800/ts):] = 0.1    # 0.1 ug/s from 1800s onward
 
 # Patient object
-George_2 = simulator.Patient([age, height, weight, gender],
-                             ts=ts,
-                             model_propo="Schnider",
-                             model_remi="Minto",
-                             model_nore="Beloeil",
-                             random_PD=False)
+George_2 = patient.Patient([age, height, weight, gender],
+                           ts=ts,
+                           model_propo="Schnider",
+                           model_remi="Minto",
+                           model_nore="Beloeil",
+                           random_PD=False)
 
 
-
-df_George_2 = George_2.full_sim(u_propo = propofol_infusion_profile, 
-                                u_remi = remifentanil_infusion_profile,
-                                u_nore = norepinephrine_infusion_profile)
+df_George_2 = George_2.full_sim(u_propo=propofol_infusion_profile,
+                                u_remi=remifentanil_infusion_profile,
+                                u_nore=norepinephrine_infusion_profile)
 
 # %% Simulation ts 50 one_step
 # Sampling time in seconds
@@ -108,25 +107,24 @@ remifentanil_infusion_profile[int(200/ts):] = 0.1    # 0.1 ug/s from 200s onward
 norepinephrine_infusion_profile[int(1800/ts):] = 0.1    # 0.1 ug/s from 1800s onward
 
 # Patient object
-George_3 = simulator.Patient([age, height, weight, gender],
-                             ts=ts,
-                             model_propo="Schnider",
-                             model_remi="Minto",
-                             model_nore="Beloeil",
-                             random_PD=False)
+George_3 = patient.Patient([age, height, weight, gender],
+                           ts=ts,
+                           model_propo="Schnider",
+                           model_remi="Minto",
+                           model_nore="Beloeil",
+                           random_PD=False)
 
 for k in range(Nsim-1):
-    
+
     uProp_k = propofol_infusion_profile[k]
     uRemi_k = remifentanil_infusion_profile[k]
     uNore_k = norepinephrine_infusion_profile[k]
-    
+
     George_3.one_step(u_propo=uProp_k,
                       u_remi=uRemi_k,
                       u_nore=uNore_k,
                       noise=False)
-    
-    
+
 
 # %% Downsample the dataframe for tests
 df_George_1_downsampled = df_George_1[df_George_1['Time'].isin(df_George_2['Time'])]
@@ -143,19 +141,19 @@ if __name__ == '__main__':
     ax[0].plot(df_George_1['Time'], df_George_1['BIS'])
     ax[0].plot(df_George_2['Time'], df_George_2['BIS'], '.')
     ax[0].plot(George_3.dataframe['Time'], George_3.dataframe['BIS'], '*')
-    
+
     ax[1].plot(df_George_1['Time'], df_George_1['MAP'])
     ax[1].plot(df_George_2['Time'], df_George_2['MAP'], '.')
     ax[1].plot(George_3.dataframe['Time'], George_3.dataframe['MAP'], '*')
-    
+
     ax[2].plot(df_George_1['Time'], df_George_1['CO'])
     ax[2].plot(df_George_2['Time'], df_George_2['CO'], '.')
     ax[2].plot(George_3.dataframe['Time'], George_3.dataframe['CO'], '*')
-    
+
     ax[3].plot(df_George_1['Time'], df_George_1['TOL'])
     ax[3].plot(df_George_2['Time'], df_George_2['TOL'], '.')
     ax[3].plot(George_3.dataframe['Time'], George_3.dataframe['TOL'], '*')
-    
+
     ax[4].plot(df_George_1['Time'], df_George_1['u_propo'])
     ax[4].plot(df_George_1['Time'], df_George_1['u_remi'])
     ax[4].plot(df_George_1['Time'], df_George_1['u_nore'])
@@ -176,8 +174,10 @@ if __name__ == '__main__':
         ax[i].grid()
     plt.ticklabel_format(style='plain')
     plt.show()
-    
+
 # %%
+
+
 def test_full_sim_results():
     """Check that the simulations results are not affected by the sampling time and by the simulation method"""
     # Check results at low concentrations
@@ -185,7 +185,7 @@ def test_full_sim_results():
     assert max(abs(map_vector-df_George_2['MAP'])) < 1
     assert max(abs(co_vector-df_George_2['CO'])) < 1
     assert max(abs(tol_vector-df_George_2['TOL'])) < 1e-1
-        
+
     assert max(abs(bis_vector-George_3.dataframe['BIS'])) < 1e-1
     assert max(abs(map_vector-George_3.dataframe['MAP'])) < 1
     assert max(abs(co_vector-George_3.dataframe['CO'])) < 1
